+++ packages/@phased/phase/lib/mix.ts
@@ -1,1 +1,0 @@
-export { default as default } from '@phased/laravel-mix';