+++ packages/@phased/phase/rollup.config.js
@@ -1,11 +1,26 @@
 import sucrase from "@rollup/plugin-sucrase";
 import commonjs from "@rollup/plugin-commonjs";
 import resolve from "@rollup/plugin-node-resolve";
-// import { terser } from "rollup-plugin-terser";
+import { terser } from "rollup-plugin-terser";
 import alias from "@rollup/plugin-alias";
 import pkg from "./package.json";
 
 const production = !process.env.ROLLUP_WATCH && process.env.NODE_ENV === 'production';
+
+const externals = {
+  node: [
+    "path",
+    "fs",
+    "child_process",
+    "os",
+    "assert",
+    "events",
+    "util",
+    "module",
+    "stream",
+    "constants"
+  ]
+}
 
 export default [
   {
@@ -15,38 +30,36 @@ export default [
       { file: pkg.module, format: "es" }
     ],
     external: [
-      "@phased/laravel-mix",
-      "@phased/state",
-      "@phased/routing",
-      "@phased/webpack-plugin",
+      // State
+      "axios",
+      "vue",
+      "vuex",
+
+      // Routing
+      "fs-extra",
+      "prettier",
+
+      // Webpack
+      'webpack',
+      'laravel-mix',
+      ...externals.node
     ],
     plugins: [
       alias({
-        entries: [
-          {
-            find: "@",
-            replacement: "./lib",
-          }
-        ],
-        customResolver: resolve({
-          extensions: ['ts']
-        })
+        entries: [ { find: "@", replacement: "./lib" } ],
+        customResolver: resolve({ extensions: ['ts'] })
       }),
 
-      resolve({
-        extensions: [".js", ".ts"]
-      }),
+      resolve({ extensions: [".js", ".ts"] }),
 
       sucrase({
         exclude: ["node_modules/**", "types/**", "__tests__"],
         transforms: ["typescript"]
       }),
 
-      commonjs({
-        //
-      }),
+      commonjs({ namedExports: { "fs-extra": ["outputFileSync"] } }),
 
-      // production && terser()
+      production && terser()
     ]
   }
 ];